requirements.txt
streamlit==1.12.0
altair<5.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
scrapingbee>=2.0.0
beautifulsoup4>=4.12.0
pandas>=1.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
plotly>=5.15.0
